Refactor: secure crud with isloggedin mware
