+++ server/Authentication_Config/routes.js
@@ -1,25 +1,5 @@
 // main authentication router
-// route middleware, the main function that checks if a user is logged in
-const isLoggedIn = (req, res, next) => {
-  // if user is authenticated in the session, carry on
-  if (req.isAuthenticated()) return next();
-
-  // if they aren't populate the profile page accordingly
-  const headerObject = req.headers;
-  // the x-forwarded-for property of the header does not appear
-  // for local host so add an alternative or will
-  // error out locally on split to get the ip address
-  // the rest of the requests are common to loacl and remote
-  let ip = (headerObject['x-forwarded-for'] || req.socket.remoteAddress).split(',')[0];
-  ip = (ip === '::1') ? 'local' : ip;
-  res.json({
-    authenticated: false,
-    userip: ip,
-    username: null,
-    displayname: null,
-  });
-  return null;
-};
+const isLoggedIn = require('./isloggedin');
 
 const authRoutes = (app, passport) => {
 // wether a user is logged in or not json data will show up on the profile page
@@ -34,7 +14,6 @@ const authRoutes = (app, passport) => {
       displayname: req.user.twitter.displayName, // only expose username and displayname
     });
   });
-  // wether a user is logged in or not json data will show up on the profile page
 
   // guest login path -- fake authentication to provide
   // semi/persistence instaed of doing it on client side
